Strengthen foreach pass-through test (#296)
Assert that the object following a bare string retains its name.

Move the independent key-order test into its structural module to keep
the condition cases below the enforced module-size limit.
