+++ src/manifest/expand_test_cases/condition_cases.rs
@@ -232,6 +232,14 @@ fn expand_foreach_non_object_entry_is_passed_through() -> Result<()> {
         "bare string entry should pass through unexpanded: {:?}",
         targets.first()
     );
+    let second_target = targets
+        .get(1)
+        .and_then(ManifestValue::as_object)
+        .context("second target object")?;
+    anyhow::ensure!(
+        second_target.get("name").and_then(ManifestValue::as_str) == Some("real"),
+        "second target should remain object named real: {second_target:?}"
+    );
     Ok(())
 }
 
@@ -295,35 +303,6 @@ targets:
         names == ["ALPHA", "BETA"],
         "expected uppercased names from Jinja filter: {names:?}"
     );
-    Ok(())
-}
-
-#[test]
-fn expand_foreach_preserves_object_key_order() -> Result<()> {
-    let env = Environment::new();
-    let yaml = r"targets:
-  - name: literal
-    vars:
-      existing: keep
-    foreach:
-      - 1
-      - 2
-    when: 'true'
-    after: done
-";
-    let mut doc: ManifestValue = serde_saphyr::from_str(yaml)?;
-    expand_foreach(&mut doc, &env)?;
-    let targets = targets(&doc)?;
-    anyhow::ensure!(targets.len() == 2, "expected expanded targets");
-    for target in targets {
-        let map = target.as_object().context("target object")?;
-        let keys: Vec<&str> = map.keys().map(String::as_str).collect();
-        anyhow::ensure!(
-            keys == ["name", "vars", "after"],
-            "key order should remain stable: {:?}",
-            keys
-        );
-    }
     Ok(())
 }
