+++ src/manifest/expand_tests.rs
@@ -10,6 +10,9 @@ mod action_condition_cases;
 #[path = "expand_test_cases/condition_cases.rs"]
 mod condition_cases;
 
+#[path = "expand_test_cases/structure_cases.rs"]
+mod structure_cases;
+
 #[path = "expand_test_cases/property_cases.rs"]
 mod property_cases;
 #[path = "expand_test_cases/target_command_available_cases.rs"]
